fix: 집 생성하기 리워드 포인트 입력시 튕김 문제 수정

+++ core/data/src/commonMain/kotlin/com/devndev/homen/core/data/model/home/request/CreateHomeRequest.kt
@@ -50,5 +50,5 @@ fun Chore.toDataModel() = ChoreRequest(
 
 fun Reward.toDataModel() = RewardRequest(
     name = name,
-    goalPoint = goalPoint
+    goalPoint = goalPoint.toInt()
 )
